Add pydantic in conda environment
