+++ .ci_support/environment-docs.yml
@@ -13,5 +13,6 @@ dependencies:
 - tqdm =4.67.3
 - jupyter-book =1.0.0
 - python =3.12
+- pydantic =2.12.5
 - hatchling =1.29.0
 - hatch-vcs =0.5.0
