+++ .ci_support/environment.yml
@@ -10,3 +10,4 @@ dependencies:
 - tqdm =4.67.3
 - hatchling =1.29.0
 - hatch-vcs =0.5.0
+- pydantic =2.12.5
